Begin rewriting existing js scripts to ts
- create tsconfig
- udpate other configs
- refactor initialization scripts to typescript
