+++ .gitignore
@@ -16,6 +16,7 @@
 
 # production
 /build
+/dist
 
 # misc
 .DS_Store
