+++ .vscode/settings.json
@@ -1,8 +1,5 @@
 {
   "editor.tabSize": 2,
-  "eslint.options": {
-    "parser": "babel-eslint"
-  },
   "debug.toolBarLocation": "docked",
   "cSpell.words": [
     "inahand"
