+++ package.json
@@ -23,7 +23,9 @@
     "yamljs": "^0.3.0"
   },
   "scripts": {
-    "start": "node server/server.js",
+    "start": "rm -rf ./dist && tsc && node dist/server.js",
+    "start:dev": "nodemon --watch server --ext js,ts --exec ts-node ./server.ts",
+    "lint": "eslint . --ext .ts,.tsx",
     "test": "mocha"
   },
   "eslintConfig": {},
@@ -34,7 +36,21 @@
     "not op_mini all"
   ],
   "devDependencies": {
+    "@babel/core": "^7.14.0",
     "@babel/eslint-parser": "^7.13.14",
+    "@types/chai": "^4.2.18",
+    "@types/cors": "^2.8.10",
+    "@types/express": "^4.17.11",
+    "@types/express-session": "^1.17.3",
+    "@types/mocha": "^8.2.2",
+    "@types/mongoose": "^5.10.5",
+    "@types/morgan": "^1.9.2",
+    "@types/sinon": "^10.0.0",
+    "@types/swagger-ui-express": "^4.1.2",
+    "@types/validator": "^13.1.3",
+    "@types/yamljs": "^0.2.31",
+    "@typescript-eslint/eslint-plugin": "^4.22.1",
+    "@typescript-eslint/parser": "^4.22.1",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "eslint": "^7.26.0",
@@ -43,10 +59,13 @@
     "eslint-plugin-import": "^2.21.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-mocha": "^8.1.0",
-    "eslint-plugin-prettier": "3.4.0",
+    "eslint-plugin-prettier": "^3.4.0",
     "mocha": "^8.4.0",
-    "prettier": ">=1.13.0",
+    "nodemon": "^2.0.7",
+    "prettier": "^2.2.1",
     "sinon": "^10.0.0",
-    "supertest": "^6.0.1"
+    "supertest": "^6.0.1",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.4"
   }
 }
